+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/sample_forms/add_ons/barcodes.jpg
