+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/sample_forms/add_ons/fonts_and_languages.png
