+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/v3.2_and_later/sample_analyze_addon_highres.py
@@ -11,10 +11,10 @@ FILE: sample_analyze_addon_highres.py
 
 DESCRIPTION:
     This sample demonstrates how to recognize documents with improved quality using
-    the add-on 'highResolution' capability.
+    the add-on 'OCR_HIGH_RESOLUTION' capability.
 
     Add-on capabilities are available within all models except for the Business card
-    model. The sample uses Layout model to demonstrate.
+    model. This sample uses Layout model to demonstrate.
 
 USAGE:
     python sample_analyze_addon_highres.py
@@ -38,7 +38,7 @@ def format_polygon(polygon):
     return ", ".join([f"[{p.x}, {p.y}]" for p in polygon])
 
 
-def analyze_layout():
+def analyze_with_highres():
     path_to_sample_documents = os.path.abspath(
         os.path.join(
             os.path.abspath(__file__),
@@ -48,7 +48,6 @@ def analyze_layout():
         )
     )
 
-    # [START extract_layout]
     from azure.core.credentials import AzureKeyCredential
     from azure.ai.formrecognizer import DocumentAnalysisClient, AnalysisFeature
 
@@ -112,14 +111,13 @@ def analyze_layout():
                 )
 
     print("----------------------------------------")
-    # [END extract_layout]
 
 
 if __name__ == "__main__":
     from azure.core.exceptions import HttpResponseError
 
     try:
-        analyze_layout()
+        analyze_with_highres()
     except HttpResponseError as error:
         print(
             "For more information about troubleshooting errors, see the following guide: "
